requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
flake8==6.0.0
idna==3.4
isort==5.11.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
